Move gatling.js into weapons/ directory
Move file and update imports
